% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_cells.R
\name{csv_cells}
\alias{csv_cells}
\title{Internal function}
\usage{
csv_cells(data_file, data_out_file)
}
\arguments{
\item{data_file_list}{csv file to convert}

\item{file_out_list}{location for rds conversion}
}
\description{
Internal function
}
