% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_handler.R
\name{csv_handler}
\alias{csv_handler}
\title{Internal function}
\usage{
csv_handler(data_file_list, file_out_list)
}
\arguments{
\item{data_file_list}{csv file to convert}

\item{file_out_list}{location for rds conversion}
}
\description{
Internal function
}
