% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_sheet_cells.R
\name{csv_sheet_cells}
\alias{csv_sheet_cells}
\title{Internal function}
\usage{
csv_sheet_cells(file_name, sheet_name)
}
\arguments{
\item{file_name}{csv file to convert}

\item{sheet_name}{ignored}
}
\description{
Internal function
}
