% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handler.R
\name{file_handler}
\alias{file_handler}
\title{Read a range of spreadsheet file types in as unformatted cells}
\usage{
file_handler(in_files, out_files)
}
\arguments{
\item{in_files}{Vector of files to read in}

\item{out_files}{Vector of file names to export to}
}
\description{
Read a range of spreadsheet file types in as unformatted cells.
}
