% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matching_sheets.R
\name{find_matching_sheets}
\alias{find_matching_sheets}
\title{Functions to check against template format}
\usage{
find_matching_sheets(gov_rds_file)
}
\arguments{
\item{gov_rds_file}{rds file to check}
}
\description{
Functions to check against template format
}
