% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_links_limited.R
\name{get_data_links_limited}
\alias{get_data_links_limited}
\title{Internal function wrapping get_data_links}
\usage{
get_data_links_limited(url_to_scrape)
}
\description{
Internal function wrapping get_data_links
}
