% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gov_file.R
\name{get_gov_file}
\alias{get_gov_file}
\title{Internal function}
\usage{
get_gov_file(file_url, dl_stem, url_stem)
}
\arguments{
\item{file_url}{url on gov.uk}

\item{dl_stem}{where to download to}

\item{url_stem}{stem for url}
}
\description{
Internal function
}
