% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gov_contents.R
\name{gov_contents}
\alias{gov_contents}
\title{Check gov.uk links for potential files}
\usage{
gov_contents(links)
}
\arguments{
\item{links}{links on gov.uk to check for files to download}
}
\description{
Takes a vector of results for pages and scrapes links to embedded data
returns a list as long as the input with:
url: original url requested
meta_data: data.table of meta data including org and date
data_titles: a vector of text titles for data files
data_links: a vector of links to data files
}
