% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gov_downloads.R
\name{gov_downloads}
\alias{gov_downloads}
\title{Download files from gov.uk}
\usage{
gov_downloads(file_list, ...)
}
\arguments{
\item{file_list}{vector of urls for files to download}

\item{dl_stem}{Local file location to use for downloaded}

\item{url_stem}{stem of the url to strip, remainder subfolders will be used when saving locally}
}
\description{
Download from gov.uk
}
