% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govuk_search.R
\name{gov_search}
\alias{gov_search}
\title{Search gov.uk for a specific term}
\usage{
gov_search(search_term = "\\"workforce management information\\"")
}
\arguments{
\item{search_term}{Phrase to search for on gov.uk}
}
\description{
Search gov.uk for a specific term and return results as a data frame
}
