% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govuk_search_json_get.R
\name{gov_search_json_get}
\alias{gov_search_json_get}
\title{Internal function to query gov.uk and parse JSON}
\usage{
gov_search_json_get(q)
}
\arguments{
\item{q}{structured query}
}
\description{
Internal function to query gov.uk and parse JSON
}
