% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_data_format.R
\name{labelled_data_format}
\alias{labelled_data_format}
\title{Format a labelled data set}
\usage{
labelled_data_format(lab_data_set)
}
\arguments{
\item{lab_data_set}{rds file to check}
}
\description{
Format a labelled data set
}
