% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter_seq.R
\name{letter_seq}
\alias{letter_seq}
\title{Letter sequence for excel columns}
\usage{
letter_seq(x)
}
\arguments{
\item{x}{Column number}
}
\description{
Generates a letter sequence that matches excel columns.
}
