% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_to_templates.R
\name{matching_to_templates}
\alias{matching_to_templates}
\title{Internal function}
\usage{
matching_to_templates(raw_sheet_data)
}
\arguments{
\item{raw_sheet_data}{check match on imported file}
}
\description{
Internal function
}
