% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_cells.R
\name{ods_cells}
\alias{ods_cells}
\title{Internal function}
\usage{
ods_cells(data_file, data_out_file)
}
\arguments{
\item{data_file_list}{ods file to convert}

\item{file_out_list}{location for rds conversion}
}
\description{
Read all sheets in an ODS file, convert to cells, save as list of tables in RDS
}
