% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_handler.R
\name{ods_handler}
\alias{ods_handler}
\title{Internal function}
\usage{
ods_handler(data_file_list, file_out_list)
}
\arguments{
\item{data_file_list}{ods file to convert}

\item{file_out_list}{location for rds conversion}
}
\description{
Internal function
}
