% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_sheet_cells.R
\name{ods_sheet_cells}
\alias{ods_sheet_cells}
\title{Internal function}
\usage{
ods_sheet_cells(file_name, sheet_name)
}
\arguments{
\item{file_name}{ods file to convert}

\item{sheet_name}{sheet names to iterate over}
}
\description{
Read a single sheet from an ODS file and convert to long form 'unpivotr' format
}
