% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_sanitiser.R
\name{text_sanitiser}
\alias{text_sanitiser}
\title{Strip line returns from text}
\usage{
text_sanitiser(x)
}
\arguments{
\item{x}{text to strip out escapes etc}
}
\description{
Strip line returns from text
}
