% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_cells.R
\name{xls_cells}
\alias{xls_cells}
\title{Internal function}
\usage{
xls_cells(data_file, data_out_file)
}
\arguments{
\item{data_file_list}{xls file to convert}

\item{file_out_list}{location for rds conversion}
}
\description{
Internal function
}
