% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_handler.R
\name{xls_handler}
\alias{xls_handler}
\title{Internal function}
\usage{
xls_handler(data_file_list, file_out_list)
}
\arguments{
\item{data_file_list}{xls file to convert}

\item{file_out_list}{location for rds conversion}
}
\description{
Internal function
}
