% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_sheet_cells.R
\name{xls_sheet_cells}
\alias{xls_sheet_cells}
\title{Internal function}
\usage{
xls_sheet_cells(file_name, sheet_name)
}
\arguments{
\item{file_name}{xls file to convert}

\item{sheet_name}{sheet names to iterate over}
}
\description{
Internal function
}
