% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gov_contents.R
\name{gov_contents}
\alias{gov_contents}
\title{Check gov.uk links for potential files}
\usage{
gov_contents(links)
}
\arguments{
\item{links}{links on gov.uk to check for files to download}
}
\description{
Check gov.uk search results for potential files to download
}
